4_busqueda_places_tipo/requirements.txt
requests==2.25.0
python_Levenshtein==0.12.2
